Move varfont/duplicate_instance_names to Universal profile
googlefonts/varfont/duplicate_instance_names from Google Fonts profile.
Renamed to varfont/duplicate_instance_names

(PR #4937)
