+++ Lib/fontbakery/legacy_checkids.py
@@ -142,7 +142,6 @@ renaming_map = {
     "com.google.fonts/check/unitsperem_strict":                             "googlefonts/unitsperem",
     "com.google.fonts/check/os2/use_typo_metrics":                          "googlefonts/use_typo_metrics",
     "com.google.fonts/check/varfont/bold_wght_coord":                       "googlefonts/varfont/bold_wght_coord",
-    "com.google.fonts/check/varfont_duplicate_instance_names":              "googlefonts/varfont/duplicate_instance_names",
     "com.google.fonts/check/varfont/generate_static":                       "googlefonts/varfont/generate_static",
     "com.google.fonts/check/varfont/has_HVAR":                              "googlefonts/varfont/has_HVAR",
     "com.google.fonts/check/vendor_id":                                     "googlefonts/vendor_id",
@@ -327,6 +326,7 @@ renaming_map = {
     "com.google.fonts/check/valid_glyphnames":                              "valid_glyphnames",
     "com.google.fonts/check/varfont/consistent_axes":                       "varfont/consistent_axes",
     "com.google.fonts/check/varfont/duplexed_axis_reflow":                  "varfont/duplexed_axis_reflow",
+    "com.google.fonts/check/varfont/duplicate_instance_names":              "varfont/duplicate_instance_names",
     "com.google.fonts/check/varfont/instances_in_order":                    "varfont/instances_in_order",
     "com.google.fonts/check/varfont/unsupported_axes":                      "varfont/unsupported_axes",
     "com.microsoft/check/vtt_volt_data":                                    "vtt_volt_data",
