+++ Lib/fontbakery/profiles/fontbureau.py
@@ -32,6 +32,7 @@ PROFILE = {
         "typographic_family_name",
         "varfont/consistent_axes",
         "varfont/duplexed_axis_reflow",
+        "varfont/duplicate_instance_names",
         "varfont/instances_in_order",
         "varfont/unsupported_axes",
         "vtt_volt_data",  # very similar to vttclean, may be a good idea to merge them.
