+++ Lib/fontbakery/profiles/googlefonts.py
@@ -143,7 +143,6 @@ PROFILE = {
             "googlefonts/unitsperem",
             "googlefonts/use_typo_metrics",
             "googlefonts/varfont/bold_wght_coord",
-            "googlefonts/varfont/duplicate_instance_names",
             "googlefonts/varfont/generate_static",
             "googlefonts/varfont/has_HVAR",
             "googlefonts/vendor_id",
