+++ Lib/fontbakery/profiles/microsoft.py
@@ -52,6 +52,7 @@ PROFILE = {
         "unwanted_aat_tables",
         "varfont/consistent_axes",
         "varfont/duplexed_axis_reflow",
+        "varfont/duplicate_instance_names",
         "varfont/instances_in_order",
         "varfont/unsupported_axes",
         "vttclean",  # very similar to vtt_volt_data, may be a good idea to merge them.
