+++ Lib/fontbakery/profiles/typenetwork.py
@@ -72,7 +72,6 @@ PROFILE = {
         "Google Fonts": [
             "googlefonts/family/equal_codepoint_coverage",
             "googlefonts/varfont/bold_wght_coord",
-            "googlefonts/varfont/duplicate_instance_names",
             "googlefonts/STAT/axis_order",
         ],
         "Outline Checks": [
