+++ docs/source/fontbakery/checks/.gitignore
@@ -1,3 +1,4 @@
+universal.rst
 adobefonts.rst
 fontbureau.rst
 fontval.rst
